@@ -1,9 +1,9 @@
 package com.zoe.framework.components.util.value.validate.annotation;
 
 import com.zoe.framework.components.util.value.validate.validator.NotContainBlankConstraintValidatorForCollection;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -22,7 +22,7 @@ import java.lang.annotation.Target;
 })
 public @interface NotContainBlank {
 
-    String message() default "{javax.validation.constraints.NotBlank.message}";
+    String message() default "{jakarta.validation.constraints.NotBlank.message}";
 
     Class<?>[] groups() default {};
 
