@@ -6,11 +6,11 @@ import com.zoe.framework.components.core.exception.common.CommonExceptionCode;
 import com.zoe.framework.components.starter.web.exception.core.interceptor.ExceptionHandlerInterceptor;
 import com.zoe.framework.components.util.value.data.CollectionUtil;
 import com.zoe.framework.components.util.value.validate.ErrorMessageUtil;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolationException;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.ConstraintViolationException;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
