@@ -2,12 +2,12 @@ package com.zoe.framework.components.util.value.validate;
 
 import cn.hutool.core.util.StrUtil;
 import com.zoe.framework.components.core.exception.defined.illegal.ParameterException;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
 import lombok.experimental.UtilityClass;
 import org.hibernate.validator.HibernateValidator;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
