@@ -2,9 +2,9 @@ package com.zoe.framework.components.util.value.validate.validator;
 
 import com.zoe.framework.components.util.value.data.StrUtil;
 import com.zoe.framework.components.util.value.validate.annotation.NotContainBlank;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
 import java.util.Collection;
 
 /**
